+++ pages/index.tsx
@@ -1,21 +1,28 @@
 import { Layout } from "@components/layout";
 import { Section } from "@components/section";
 import { CustomImage } from "@components/image";
-import { NextPage } from "next";
+import { GetStaticProps, NextPage } from "next";
+import { getPortfolioProjects } from "@util/notion";
+import { IProject } from "@util/interface";
+import { ProjectPanel, TechStack } from "@components/panel";
 
-const Home: NextPage = () => {
+type HomeProps = {
+    projects: IProject[];
+};
+
+const Home: NextPage<HomeProps> = ({ projects }: HomeProps) => {
     return (
         <Layout title="Home" description="2022 Portfolio">
             <Section className="justify-center min-h-screen">
-                <div className="flex flex-col">
-                    <div className="flex-grow flex flex-col m-auto">
-                        <span className="font-clash text-6xl lg:text-9xl font-medium mx-auto">
+                <div className="flex w-1/2 flex-col m-auto">
+                    <div className="flex-grow flex flex-col">
+                        <span className="font-clash text-6xl lg:text-9xl font-medium">
                             Alan John
                         </span>
-                        <span className="font-clash text-6xl lg:text-9xl font-light mx-auto">
+                        <span className="font-clash text-6xl lg:text-9xl font-light self-center">
                             Software
                         </span>
-                        <span className="font-clash text-6xl lg:text-9xl font-light mx-auto -mt-2 lg:-mt-8">
+                        <span className="font-clash text-6xl lg:text-9xl font-light self-center -mt-2 lg:-mt-8">
                             Engineer
                         </span>
                     </div>
@@ -62,8 +69,30 @@ const Home: NextPage = () => {
                     </div>
                 </div>
             </Section>
+            <Section title="skills" className="mt-16">
+                <div className="flex flex-col lg:flex-row lg:columns-3 w-11/12 mx-auto">
+                    <div className="flex lg:w-2/5 lg:order-1">
+                        <ProjectPanel projects={projects} />
+                    </div>
+                    <div className="order-first lg:order-2 lg:w-1/5 flex">
+                        <span className="m-auto">Code</span>
+                    </div>
+                    <div className="flex lg:w-2/5 lg:order-3">
+                        <TechStack />
+                    </div>
+                </div>
+            </Section>
         </Layout>
     );
+};
+
+export const getStaticProps: GetStaticProps<HomeProps> = async () => {
+    const response = await getPortfolioProjects();
+    return {
+        props: {
+            projects: response,
+        },
+    };
 };
 
 export default Home;