+++ src/components/image/index.tsx
@@ -11,12 +11,5 @@ export const CustomImage: React.FC<ImageProps> = ({
     height,
     width,
 }: ImageProps) => {
-    return (
-        <Image
-            src={src}
-            height={height}
-            width={width}
-            className="grayscale hover:grayscale-0 transition-all duration-300"
-        />
-    );
+    return <Image src={src} height={height} width={width} className="" />;
 };