+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2022-01-29
+### Added
+- Panels to show projects (unstyled)
+- Notion support to read Project database
+### Fixed
+- Tag component padding
+- Round button styling
+- Home hero component alinment
+### Removed
+- Image grayscale filter
+
 ## [0.3.1] - 2022-01-28
 ### Added 
 - Sharp for image optimization